Another major refactor to ensure package structure and namespacing
