+++ data/user_data.json
@@ -1,1 +1,1 @@
-{"Nate": 100, "Checkers Master": 250, "Pey": 150, "wash": 0, "awshi": 0}
+{"Nate": 100, "Checkers Master": 250, "Pey": 150, "wash": 0, "awshi": 0, "washi": 0}
