+++ src/checkers/core/constants.py
