+++ src/checkers/core/player.py
