+++ src/checkers/ui/menu.py
